pyproject.toml
[project]
name = "DividirPodcast"
version = "0.1.0"
description = ""
authors = ["Rodrigo Meléndez <rodmlndz.com>"]
requires-python = ">=3.11"
dependencies = [
    "ffmpeg==1.4",
    "ffprobe==0.5",
    "python3-openid==3.2.0",
    "python-dotenv",
    "flask>=3.1.1,<4",
    "requests>=2.31.0,<3",
    "supabase>=2.0.0"
]